back/requirements.txt
fastapi
uvicorn[standard]
chromadb==1.0.20
sentence-transformers
pydantic-settings
python-dotenv
pymupdf
python-docx
openai>=1.40.0
numpy